Rubocop fixes
Ref: https://github.com/hl7au/au-ps-inferno/issues/43
