+++ .rubocop.yml
@@ -5,7 +5,6 @@ AllCops:
     - "lib/au_ps_inferno/0.5.0-preview/**/*"
     - "lib/au_ps_inferno/1.0.0-ballot/**/*"
     - "aidbox_config.rb"
-    - "lib/au_ps_inferno/utils/composition_utils.rb"
     - "lib/au_ps_inferno/generator/test_config_registry.rb"
     - "lib/au_ps_inferno/generator/suite_structure.rb"
     - "lib/au_ps_inferno/utils/composition_decorator.rb"
@@ -13,8 +12,6 @@ AllCops:
     - "lib/au_ps_inferno/generator/retrieve_cs_group_generator.rb"
     - "lib/au_ps_inferno/generator/metadata_manager.rb"
     - "lib/au_ps_inferno/generator/generator.rb"
-    - "lib/au_ps_inferno/generator/retrieve_cs_group_tests.rb"
-    - "lib/au_ps_inferno/utils/section_names_mapping.rb"
 
 Gemspec/DevelopmentDependencies:
   Enabled: false
