+++ lib/au_ps_inferno/utils/composition_utils.rb
@@ -1,9 +1,14 @@
 # frozen_string_literal: true
 
 require_relative 'bundle_decorator'
+require_relative 'composition_utils/ms_elements'
+require_relative 'composition_utils/boolean_and_stats'
 
 # Utilities for FHIR Composition resources
 module CompositionUtils
+  include CompositionUtilsMsElements
+  include CompositionUtilsBooleanAndStats
+
   def scratch_bundle
     scratch[:bundle_ips_resource]
   end
@@ -56,119 +61,5 @@ module CompositionUtils
     else
       warning "Section #{section.code.coding.first.display}(#{section_code}) has no entries"
     end
-  end
-
-  def boolean_to_existent_string(boolean_value)
-    boolean_value ? '✅ Passed' : '❌ Failed'
-  end
-
-  def boolean_to_existent_string(boolean_value)
-    boolean_value ? '✅ Populated' : '❌ Missing'
-  end
-
-  def execute_statistics(resource, path_expression, humanized_name)
-    data_value = resolve_path(resource, path_expression).first.present?
-    boolean_value = boolean_to_existent_string(data_value)
-    "**#{humanized_name}**: #{boolean_value}"
-  end
-
-  def all_entries_have_full_url_info?
-    entry_full_url_count = resolve_path(scratch_bundle, 'entry.fullUrl').length
-    entries_count = resolve_path(scratch_bundle, 'entry').length
-
-    entry_full_url_count == entries_count
-  end
-
-  def timestamp_info?
-    resolve_path(scratch_bundle, 'timestamp').first.present?
-  end
-
-  def type_info?
-    resolve_path(scratch_bundle, 'type').first.present?
-  end
-
-  def identifier_info?
-    resolve_path(scratch_bundle, 'identifier').first.present?
-  end
-
-  def check_section_element_completeness(path_expression)
-    sections_count = resolve_path(composition_resource, path_expression).length
-    selected_by_expression_count = resolve_path(composition_resource, path_expression).length
-
-    boolean_to_existent_string(sections_count == selected_by_expression_count)
-  end
-
-  def composition_mandatory_ms_elements_info(
-    optional_ms_elements,
-    mandatory_ms_elements,
-    optional_ms_sub_elements,
-    mandatory_ms_sub_elements,
-    mandatory_ms_slices,
-    optional_ms_slices
-  )
-    check_bundle_exists_in_scratch
-    check_mandatory_ms_elements(mandatory_ms_elements)
-    check_optional_ms_elements(optional_ms_elements)
-    check_mandatory_ms_sub_elements(mandatory_ms_sub_elements)
-    check_optional_ms_sub_elements(optional_ms_sub_elements)
-    check_mandatory_ms_slices(mandatory_ms_slices)
-    check_optional_ms_slices(optional_ms_slices)
-
-    mandatory_ms_elements_passed = all_elements_passed?(mandatory_ms_elements + mandatory_ms_sub_elements)
-    assert mandatory_ms_elements_passed, 'Mandatory Must Support elements are not populated'
-  end
-
-  def check_mandatory_ms_slices(mandatory_ms_slices)
-    info_block('Mandatory Must Support slices SHALL be correctly populated if a value is known', mandatory_ms_slices)
-  end
-
-  def check_optional_ms_slices(optional_ms_slices)
-    info_block('Optional Must Support slices SHALL be correctly populated if a value is known', optional_ms_slices)
-  end
-
-  def info_block(title, elements)
-    info "**#{title}**:\n\n#{composition_mandatory_elements_info(elements)}"
-  end
-
-  def check_mandatory_ms_elements(mandatory_ms_elements)
-    info_block('Mandatory Must Support elements are correctly populated',
-               mandatory_ms_elements)
-  end
-
-  def check_optional_ms_elements(optional_ms_elements)
-    info_block(
-      'Optional Must Support elements are correctly populated', optional_ms_elements
-    )
-  end
-
-  def check_mandatory_ms_sub_elements(mandatory_ms_sub_elements)
-    info_block(
-      'Mandatory Must Support sub-elements of a complex element SHALL be correctly populated if a value is known', mandatory_ms_sub_elements
-    )
-  end
-
-  def check_optional_ms_sub_elements(optional_ms_sub_elements)
-    info_block(
-      'Optional Must Support sub-elements of a complex element SHALL be correctly populated if a value is known', optional_ms_sub_elements
-    )
-  end
-
-  # Generic: renders pass/fail for each element (or sub-element) in the given list.
-  # Do not append section.title/section.text here; they are sub-elements and only
-  # appear when in the metadata list (e.g. composition_mandatory_ms_sub_elements).
-  def composition_mandatory_elements_info(mandatory_ms_elements)
-    mandatory_ms_elements.map do |element|
-      execute_statistics(composition_resource, element[:expression], element[:label])
-    end.join("\n\n")
-  end
-
-  def all_elements_passed?(elements)
-    elements.map do |element|
-      resolve_path(composition_resource, element[:expression]).first.present?
-    end.all?
-  end
-
-  def composition_resource
-    BundleDecorator.new(scratch_bundle).composition_resource
   end
 end
